+++ pyjsonrpc/tools.py
@@ -78,7 +78,7 @@ def safe_unicode(value):
             return unicode(value, "utf-8")
         except UnicodeDecodeError:
             return unicode(value, "iso-8859-15", "ignore")
-        except StandardError, err:
+        except StandardError as err:
             return unicode(repr(value))
-    except StandardError, err:
+    except StandardError as err:
         return unicode(repr(value))